facade refactor
