+++ Structural/Decorator/src/index.ts
