+++ Structural/Facade/src/interfaces/Calculator.ts
@@ -1,0 +1,3 @@
+export interface Calculator {
+  calculate(num1: number, num2: number): number;
+}
