+++ Structural/Facade/src/models/Add.ts
@@ -1,0 +1,7 @@
+import { Calculator } from "../interfaces/Calculator";
+
+export class Add implements Calculator {
+  calculate(num1: number, num2: number) {
+    return num1 + num2;
+  }
+}
